requirements.txt
pandas
numpy
scipy
matplotlib
apache-airflow
openpyxl
sqlalchemy
psycopg2-binary
